Reimplement gum_process_enumerate_modules
By reverse engineering dl_iterate_phdr in QNX 6.6 and dlopen in
ldqnx.so.2, determined how best to extract the module file names.

Currently a combination of /proc/self/as devctl queries and the
dlopen/dl_iterate_phdr hack is used as the former provides module
size which the latter does not.