Remove V8 from libfrida-gum-1.0.la
We really want to avoid V8 symbols ending up in the static library, and
just rely on the pkg-config to link in V8's static library. Otherwise we
run into issues when frida-gum is linked into a library that itself
depends on another version of V8, which is the case for a Node.js binding.

This is obviously a hack, and the longer term plan is to move all of
gum/gumscript* out into either its own library, or frida-core/lib/agent.