Add support for source-maps in the JavaScriptCore runtime
We also convert the stack-trace to the same format as used by V8.
Note that JavaScriptCore doesn't provide a stack-trace preparation hook
like V8 does with Error.prepareStackTrace(), so we expose a helper as
Script.symbolicate() that we call automatically for unhandled exceptions,
and allow this to be called explicitly by scripts that would like to
handle and report exceptions themselves. It is also safe to call this
from scripts running in the V8 runtime – it doesn't modify the value at
all in that case.