Add CPSR to GumCpuContext on ARM
This resolves the issue we were seeing on linux-arm where we would get
a second unexpected SIGSEGV while processing a signal in the exceptor code.

This manifested as an unexpected SIGSEGV and the r0 register being overwritten
to 0.

The root cause of the issue was that we were not correctly setting the CPSR
register in the context, resulting in the kernel's `valid_user_regs` call
failing, causing us to fall into the 'badframe' label of sys_rt_sigreturn.