Fix thread enumeration on Linux
It seems like the only reliable way to read/write the registers of
another thread is to use ptrace(). We used to accomplish this by
hi-jacking the target thread by installing a signal handler and sending a
real-time signal directed at the target thread, and thus relying on the
signal handler getting called in that thread. The signal handler would
then provide us with read/write access to its registers. This hack would
however not work if a thread was for example blocking in poll(), as the
signal would then just get queued and we'd end up waiting indefinitely.

It is however not possible to ptrace() another thread when we're in the
same process group. This used to be supported in old kernels, but it was
buggy and eventually dropped. So in order to use ptrace() we will need to
spawn a new thread in a different process group so that it can ptrace()
the target thread inside our process group. This is also the solution
recommended by Linus:

https://lkml.org/lkml/2006/9/1/217

Because libc implementations don't expose an API to do this, and the
thread setup code is private, where the TLS part is crucial for even just
the syscall wrappers - due to them accessing `errno` - we cannot make any
libc calls in this thread. And because the libc's clone() syscall wrapper
typically writes to the child thread's TLS structures, which we cannot
portably set up correctly, we cannot use the libc clone() syscall wrapper
either.