Make the Duktape runtime preemptible
This means we can let go of our JavaScript lock while for example calling
out to a potentially long-running `NativeFunction`, and avoids dead-lock
situations where the called function needs a lock that another thread is
holding while trying to get the JS lock (e.g. because of an intercepted
function).

Our V8 runtime already behaves this way, and with these changes plus our
Duktape patch we finally have the same behavior on the Duktape runtime.