Add SystemFunction API
One shortcoming of NativeFunction is that it is not possible to get
`errno` / `GetLastError()` reliably, as Frida may have clobbered it.

This new API is just like NativeFunction except that the return value
is an object containing `value` and `errno` (or `lastError` on Windows).

For example calling `open()` with a path that does not exist, the return
value on Darwin would be:

    { value: -1, errno: 2 /* ENOENT */ }