Fix crash when processing tick callbacks
We were using the heap thread regardless of which Duktape thread we were
about to leave. This resulted in a fatal error when more than one native
thread was running JS in lockstep, and any of the non-heap threads would
process tick callbacks. (We recycle the heap thread and only create new
Duktape threads when another native thread is using the heap thread.)