Fix use-after-free caused by race condition on unload
We need to hold a strong reference to the script while processing the
on_unload callbacks, as one of them might drop the last reference. This
would happen if our thread got unscheduled immediately after calling the
callback, and the other thread got a chance to process the idle handler
and let go of the last reference.