Fix performance issue with source maps
A typical source map takes ~5 seconds to parse with the Duktape runtime
on an iPhone 5S. This is way too slow, and there isn't much room for
optimizing the algorithm due to the sheer amount of bytes that need to
be munged. To solve this we implement a minimal source map parser in C
that only covers the bits that we need.