Simplify setjmp() logic and fix race-condition
Up until now we've tried very hard to hide all the gory details of
setjmp() behind the Exceptor API, but in the end it proved to be more
trouble than it's worth. By exposing these internals to the public
header it means we leak a few implementation details, but on the other
hand it makes Gum easier to port and maintain, and gets rid of one
memory allocation.

We now also protect the `scopes` data-structure against concurrent
modification (oops!).