Fix Interceptor thread context cleanup
We cannot clean it up when we get the TLS callback, as the application
might call an intercepted function in one of its TLS callbacks called
after ours. Instead we emit “flush-suggested” and defer cleanup until
a future gum_interceptor_flush(). At that point we check the lifetime
beacon and potentially perform the cleanup.