Implement side-stack and speed things up
The side-stack speeds up handling of ret-instrutions, and will also
allow us to implement the `depth` field of CALL events, like we do on
x86.

To speed things up even further we now implement the same backpatching
tricks that we do on x86, plus, inline caches for dynamic branches where
the target is in a register.