Fix unicode-handling in Memory.readCString()
As this API is intended for plain ASCII and potentially invalid UTF-8,
we need to replace bytes that cannot be interpreted as valid Unicode.

We lean on g_utf8_make_valid() to do this for us, which replaces them
with the Unicode replacement character (U+FFFD).

This is important to ensure that invalid strings never enter the
JavaScript runtimes and result in a crash when later handed to send().