Adjust size of mapped ranges (darwin)
The tail of file-mapped segments is guaranteed to be accessible only within “vm_kernel_page_size” boundary after the end of file, after that point SIGBUS is triggered - regardless of the real size of the memory map.

The returned “size” for ranges crossing the end of mapped files is now adjusted to be within the safe range.

As a plus, “file” object has the new “size” field which holds the size of the mapped file on disk in bytes.

Added the “file” object to “Module.enumerateRanges()”  results too.

__PAGEZERO segment is now skipped while enumerating Module’s ranges.