Adjust size of mapped ranges on i/macOS (#262)
The tail of file-mapped segments is only guaranteed to accessible within
the next `vm_kernel_page_size` boundary after the end of file. Accesses
beyond that point result in SIGBUS, regardless of the real size of the
memory map.

The returned `size` for ranges crossing the end of mapped files is
now adjusted to be within the safe range.

As a plus, Gum.FileMapping now has a new `size` field which holds the
size of the mapped file on disk in bytes.

This also adds `file` details to ranges emitted by Module.enumerate_ranges().

Lastly, the `__PAGEZERO` segment is now skipped while enumerating module
ranges.