Potentially cloak more than one range for threads
Implemented only on i/macOS for now.

The default behaviour when apple’s libpthread allocates resources for a thread is to create a contiguous range made of guard + stack + pthread. There’s also the possiblity to create a thread providing a separated stack, in that case no guard page will be provided by default and only the pthread structure will be allocated internally.

In any case, both the internal allocation extents and the stack itself can be retrieved by picking the right values out of the opaque pthread structure.

The `gum_thread_try_get_ranges` function detects if the allocation is contiguous or not. If it’s contiguous only one range is filled, otherwise there will be one range for stack and one for pthread structure. The number of produced ranges is returned. In backends which don’t implement this logic yet, a number of 0 ranges is returned.

Offsets in pthread structure have been pretty stable so far, the most notable change is that at some point (around iOS 10.x) the `thread_id` field has been moved after the fields we use. To account for this and provide compatibility with older versions of pthread structure, we check if `thread_id` is at the newer offset, if it’s not a skew of 8 bytes is added to all other offsets.