Potentially cloak more than one range for threads (#268)
Implemented only on i/macOS for now.

The default behavior when Apple's libpthread allocates resources for a
thread is to create a contiguous range made of guard + stack + pthread.
It's also possible to create a thread with a user-defined stack, in
which case no guard page will be provided by default and only the
pthread structure gets allocated internally.

In any case, both the internal allocation extents and the stack itself
can be retrieved by picking the right values out of the opaque pthread
structure.

The `gum_thread_try_get_ranges()` function detects if the allocation is
contiguous or not. If it's contiguous only one range is filled,
otherwise there will be one range for the stack and one for the pthread
structure. The number of produced ranges is returned. In backends that
don't yet implement this logic, a number of 0 ranges is returned.

Offsets in Apple's pthread structure have been pretty stable so far, the
most notable change is that at some point (around iOS 10.x) the
`thread_id` field got moved after the fields we use. To account for this
and provide compatibility with older versions of the pthread structure,
we check if `thread_id` is at the newer offset, if it’s not a skew of
8 bytes is added to all other offsets.