Re-generate Duktape without the Symbol builtin
Polyfill modules like es-abstract may assume that if `Symbol` exists and
`Symbol.iterator` is present, then `Array.prototype`, `String.prototype`,
and such must have it, and calls it without checking. (Kudos to @mrmacete
for reporting and digging into this!)

So to keep things simple we disable this feature, as applications needing
this kind of features are better off using frida-compile, or using our
V8-based runtime.

The Duktape snapshot was generated like this:

    python util/dist.py
    cd dist
    python tools/configure.py \
        --source-directory src-input \
        --output-directory src-frida \
        --config-metadata config \
        -DDUK_USE_DEBUGGER_DUMPHEAP \
        -DDUK_USE_DEBUGGER_INSPECT \
        -DDUK_USE_DEBUGGER_PAUSE_UNCAUGHT \
        -DDUK_USE_DEBUGGER_SUPPORT \
        -DDUK_USE_INTERRUPT_COUNTER \
        -UDUK_USE_SYMBOL_BUILTIN