Stalker: fix uaf and allow unfollow when transforming
Actually, there is no way to unfollow a stalked thread which uses transform.
Calling unfollow on the same thread will trigger an uaf which crash frida (it attempt to use a freed writer)
Calling unfollow on a different thread will trigger another unknown bug which is gonna be investigated later if needed but this, will open one little door to detach a stalked thread.
From my understanding, we should just unfollow when it's safe to do it. There is also missing logic when unfollow is triggered from a callout in example, which trigger another bug