Improve NativeFunction to support variadic functions (#369)
Such a variadic NativeFunction can be declared as e.g.:

    const func = new NativeFunction(ptr, 'int', ['pointer', '...',
        'pointer', 'float']);

Which means the variadic argument types can be `'pointer', 'float',
'pointer', 'float', ...` repeatedly. And the invocation of a variadic
function can be with any number of arguments which is greater or equal
the number of fixed arguments, such as `func(ptr(1))` or
`func(ptr(1), ptr(2))`.