Accept ArrayBuffer parameters in NativeFunction
To avoid needing to allocate and copy to memory when dealing with an
already created ArrayBuffer, e.g. returned by frida-fs.

This is a very dangerous feature as the called function may hold on to
the pointer and use it after returning. But users already have to take
this into account when using Memory.alloc().

Because of how dangerous this is we don't allow passing an ArrayBuffer
anywhere a NativePointer is expected. We limit it to NativeFunction
parameters, as we know the JS object won't get GCed while the function
call is in progress.

Fixes #320.