Ensure script context is destroyed
Even if the script is still unloaded.

This ensures that ultimately the script core is disposed, which in turn
disposes its reference to the exceptor.

Failing to do so caused indefinite hang when attaching after detaching
in presence of unloaded scripts, for the exceptor thread remaining in
the target process.