Move InspectorServer into its own library on Windows
As frida-core needs it, and we don't want the C++ object files tipping
off the linker that their C runtimes are different (static vs dynamic).

Our only component that needs to target the dynamic CRT is frida-clr,
and so far we have been able to get away with building frida-core and
its static library peers just once, targeting the static CRT but
optionally linking against the dynamic CRT in the case of frida-clr.
(Our prebuilt deps are built for both, though.)