Fix Stalker.follow() of thread in syscall on Windows
The other platforms allow us to restart the system call, but that's not
a thing on Windows (unless there's API I'm not aware of). So we handle
it by setting a hardware breakpoint on the RET instruction where the
syscall will return. This also works if it has just returned and the
instruction pointer is on the RET instruction for that reason.

This is a long-standing issue that's been causing random test failures.