Drop support for parsing the dyld shared cache
It was only used by Darwin.Mapper when the module to be mapped depended
on a module that wasn't already loaded, *and* that module happened to be
part of the dyld shared cache. This was however never needed for our own
use-cases, and remained a half-baked feature that only added friction.

So now that Apple has evolved the on-disk format even further, it's a
good time to get rid of this code.