Support linking against system GLib and libffi
Frida makes a number of modifications to GLib, GObject, and GIO in order
to function correctly across all supported platforms. This commit checks
for Frida-specific support before performing these operations, thus
allowing Frida to be built against system libraries (which typically do
not have the Frida patches applied on top of them). However, this
results in a degraded experience and thus we have to disable a number of
tests in order for the suite to pass.