Stalker/x86: Also emit call probe on jmp insns
Call probes do not work on imported functions called through the GOT, as
these are indirect jumps with the target taken from the PLT.

Also considering direct jumps as candidates for call probes additionally
allows for call probes to be triggered when called as tail-calls.

This has a performance impact at compile-time on the stalker, but should
have no performance impact at run-time.