Fix conditional Interceptor unignore logic in GumJS
We should avoid the ignore level ever becoming negative, as it means
Interceptor's internal ignore logic won't work, and listeners might end
up getting invoked when they shouldn't. Even if this could be handled
internally, it's not very elegant.

This basically means we still support the old behavior for now, but it
would make sense to later turn the ignore level into a uint, and
consider it an API contract violation to unignore() when not ignored.