Rework i/macOS Exceptor to support latest iOS 14
It is no longer possible to forward Mach exceptions to the previous
handler due to guarded Mach ports. On latest iOS 14 this results in
EXC_GUARD w/ subtype GUARD_TYPE_MACH_PORT due to ILLEGAL_MOVE.

This meant that our exception handler thread got stopped when trying to
send a message to the previous handler, generating another exception
added to our Mach exception handler queue, and hence a deadlock.

So it seems the best we can do is to revert to the previous handler,
pretend we handled the exception, and assume that an identical exception
will be generated right after. That time around the original handler
will receive the exception and be able to handle it. This also means we
won't handle exceptions past that point, at least not until we introduce
API to “unpause” exception handling.