gumjs: ignore Interceptor context in unrelated NativeCallback invokes
In this way invalid Interceptor contexts from higher up the call stack
will be safely ignored in favor of the minimal but correct callback
context.

NativeCallback instances invoked as part of `Interceptor.replace()` will
continue to behave as usual and get the full context from the
Interceptor's thread storage.

The new caveat is that when using the same NativeCallback instance both
for `Interceptor.replace()` and an Interceptor-unrelated scenario (for
example `ObjC.implement()`), AND both scenarios are invoked as part of
the same call stack so that the Interceptor case happens earlier, the
Interceptor context will "win" in the second invocation too and there
will contain possibly invalid values. To overcome this issue, it is now
possible to just use two different NativeCallback instances (which can
share the same javascript function body, if needed).