objc-api-resolver: Protect against objc_disposeClassPair via mutex
`GumObjcDisposeClassPairMonitor` uses the `Interceptor` to attach to
`objc_disposeClassPair` so that on enter it locks a mutex, releasing it
on leave.

This mutex is then used to protect the creation of the initial snapshot
of Objective-C classes, and a cache-less reimplementation of
`gum_objc_api_resolver_find_method_by_address`.

The monitor instance is shared across all concurrent usages, via the
"obtain" semantics.