gumjs: prevent File and SQLite apis from triggering interceptor
When these APIs are called from threads not ignored by the interceptor,
they can trigger hooks on some libc functions: this is not expected and
could cause assertions or undefined behaviour depending on what the
hooks do.

This change makes sure that while all calls to sqlite3 and FILE
functions are performed by gumjs, Interceptor is ignoring the thread.