darwin(x86/arm64): Initialize TLV descriptors
Within a module's image, before loading, the TLV descriptors thunk
function is set to _tlv_bootstrap, which calls abort at runtime.
The key values are also set to 0 on non-dyld_shared_cache images.

We create a new pthread_key at runtime, and store it in each TLV
descriptor. Additionally, we fix up the thunk function to tlv_get_addr,
which returns a pointer to the variable's storage as
((void**) gsbase)[descriptor->key] + descriptor->offset.

We're not handling any non-zero keys in the image, as those are only
used within the dyld_shared_cache.