gumjs: Add NativePointer#readVolatile()
For when memory is being read in a moment where another thread may
potentially unmap it or change its protection.

Since we're using gum_memory_read(), which results in a system call, we
don't have to worry about OS-exceptions on invalid memory addresses.

This makes this API particularly useful on newer versions of iOS, where
Exceptor may not be functional.