stalker: Improve exclusive store handling on arm64
Instead of potentially expanding the current block to include
instructions beyond where the block would naturally end, we move to a
safer approach: Once we encounter an exclusive store, we look back at
the previously generated blocks to see if we can find one with an
exclusive load. If we do, we mark this range of blocks as using
exclusive access. We also invalidate the blocks, so that problematic
instrumentation can be omitted upon recompilation.

To also allow custom transformers to adapt their generated code, we
introduce StalkerIterator.get_memory_access().

Co-authored-by: Håvard Sørbø <havard@hsorbo.no>