stalker: Fix crash on invalidation in added slabs
When we run out of space in the initial slab and an additional slab is
added, we need to make sure we propagate the invalidator function
pointer to it.

This was a long-standing bug that finally turned highly reproducible
with the recent changes to exclusive store handling, where invalidations
may now occur without the user explicitly requesting them.