process: Add get_main_module() (#758)
This returns the details of the module which represents the main
executable of the process.

It is exposed to JS via the `Process.mainModule` property.

For Darwin this is implemented by getting the first loaded module for
which the Mach-O header's `filetype` is set to `MH_EXECUTE`. For all
other OSes it is currently just the first loaded module (equivalent to
Process.enumerateModules()[0]).