Add Memory.queryProtection()
This change adds `gum_memory_query_protection` which takes an address
and returns the current protection for the page which contains it.

It is accessible from js bindings through `Memory.queryProtection()`.

On the Darwin backend it's a thin wrapper on top of
`gum_darwin_query_protection`, on all other platforms it relies instead
on the respective private implementations of `gum_memory_get_protection`.