memory: Make regex patterns raw
According to the docs of `g_regex_new`: "Usually strings must be valid
UTF-8 strings, using this flag they are considered as a raw sequence
of bytes."

The "strings" this refers to are the haystacks we then pass to
`g_regex_match_full` when searching.

Without the flag, memory search with regex patterns can get
interrupted before the range is over, when "invalid" bytes are
encountered, resulting in false negatives.