gumjs: Relinquish JS lock while unreffing modules
To avoid deadlocking in case dispose() releases a cached handle. Such an
operation typically requires acquiring a runtime linker lock. Another
thread might already be holding that lock while waiting for the JS lock.
A common scenario for that to happen is that the agent registers a
callback with the runtime linker, called whenever a module is loaded or
unloaded.

Kudos to @mrmacete for reporting.