gumjs: Fix crash in Module finalizers
Our QuickJS suspend/resume patch isn't elaborate enough to support usage
from finalizers. Given that modules are created and destroyed in waves,
however, it's probably a bit on the expensive side to suspend/resume JS
execution during the destruction of each value. So to avoid both issues
we defer the unref using an idle source.

The better longer term solution will probably be to introduce a
ModuleObserver of sorts that integrates with the platform's runtime
loader and manages the lifetime of each Module. This will also allow us
to emit signals anytime a Module is added/removed, which is a feature
that a lot of agents would benefit from.

Kudos to @mrmacete for reporting.