interceptor: Refactor to use memory patch code
This is done by introducing the new `gum_memory_patch_code_pages`
function, which can be used from both `gum_memory_patch_code` and the
Interceptor's transaction.

This function abstracts all page-level logic needed to make the code
patch possible, so it can function as a single point of maintenance.