darwin-module: Swap-hook lldb image notifier
Since on modern iOS devices the `lldb_image_notifier` function is only
1-instruction long, Interceptor currently fails to attach it. When
this happens, the module registry is left non-functional.

This change tries first to swap the pointer in the dyld "all image
infos" structure with our own, so there's no need to use Interceptor
in order to get image load/unload events.

On older systems where this can't be done (< iOS 4.0 / Mac OS X 10.7)
the previous logic is used as a fallback.