gumjs: Ensure consistent Interceptor onLeave call (#1058)
Avoid calling `onLeave` when `onEnter` is present but hasn't been
called yet. This can happen when attaching the listeners mid-call.

In that scenario, calling `onLeave` with a fresh invocation context
could result in unexpectedly uninitialised properties leading to
unpredictable behaviour.

This change guarantees that when both `onEnter` and `onLeave` are
present, `onLeave` is only called back after `onEnter` actually
happened.