darwin-apiresolver: Rewrite find_method_by_address
This change brings down the complexity of looking up Objective-C
methods by address, by looking up only among classes defined by the
module the address belongs to, or on which that module defines
categories on.

This allows us to find results in less 10% of the time, missing only
methods which are swizzled dynamically.